# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Auto Ecole documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permis": {
            "get": {
                "tags": [
                    "Permis"
                ],
                "summary": "Get a list of permis",
                "operationId": "5bc18a1dc63281289a67c007301720c9",
                "responses": {
                    "200": {
                        "description": "List of permis"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permis"
                ],
                "summary": "POST to create permis",
                "operationId": "c34e8731796ead06f661b956d12d7b93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/permis/{id}": {
            "put": {
                "tags": [
                    "Permis"
                ],
                "summary": "Update an permis",
                "operationId": "ab76ec16207524bfd707f9d8e37cdb48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permis to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permis updated successfully"
                    },
                    "404": {
                        "description": "Permis not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Permis"
                ],
                "summary": "Delete an permis",
                "operationId": "2aff488ee41f79df34872549223717ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permis to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "permis deleted successfully"
                    },
                    "404": {
                        "description": "permis not found"
                    }
                }
            }
        },
        "/api/personnels": {
            "get": {
                "tags": [
                    "personels"
                ],
                "summary": "Get a list of personnels",
                "operationId": "05fefd580bf90338727d99ad711e626f",
                "responses": {
                    "200": {
                        "description": "List of personnels"
                    }
                }
            },
            "post": {
                "tags": [
                    "personels"
                ],
                "summary": "POST to create personnels",
                "operationId": "182c936cb53bcd8c33df111f82cc1d41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create personnels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/personnels/{id}": {
            "put": {
                "tags": [
                    "personels"
                ],
                "summary": "Update an personnels",
                "operationId": "3026b9166ba6348fdce0c3ea806926b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the personnels to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "agence_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personnels updated successfully"
                    },
                    "404": {
                        "description": "Personnels not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "personels"
                ],
                "summary": "Delete an personnels",
                "operationId": "17df0e78d721cac90b5fd9b48e95b173",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the personnels to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "personnels deleted successfully"
                    },
                    "404": {
                        "description": "personnels not found"
                    }
                }
            }
        },
...
    }
}